Move thumbnail to common folder
