+++ libs/search-widget/src/_video-widget/ThumbnailPlayer.svelte
@@ -1,5 +1,5 @@
 <script>
-  import Thumbnail from '../widgets/row/Thumbnail.svelte';
+  import Thumbnail from '../common/thumbnail/Thumbnail.svelte';
   import { createEventDispatcher } from 'svelte';
   import Icon from '../common/icons/Icon.svelte';
   import { fade } from 'svelte/transition';
