+++ libs/search-widget/src/common/thumbnail/Thumbnail.scss
